+++ src/main/java/com/example/block/repository/TransactionReviewRepository.java
@@ -5,6 +5,6 @@ import com.example.block.domain.mapping.Review;
 import com.example.block.domain.mapping.TransactionReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionReviewRepository extends JpaRepository<TransactionReview, Long> {
+public interface TransactionReviewRepository extends JpaRepository<TransactionReview, Integer> {
     TransactionReview findAllByUserAndReview(User user, Review review);
 }
