+++ src/main/java/com/example/block/repository/UserRepository.java
@@ -1,7 +1,9 @@
 package com.example.block.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-import com.example.block.domain.User;
 
-public interface UserRepository extends JpaRepository<User, Long>{
+import com.example.block.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
+
 }
