+++ src/main/java/com/example/block/service/PointService.java
@@ -7,12 +7,12 @@ import com.example.block.domain.PointDetail;
 import java.util.List;
 
 public interface PointService {
-    public PointDetail chargePoint(Long userId, Long point);
-    public PointDetail usePoint(Long userId, Long point);
-    public Long getMyPoint(Long userId);
-    public List<PointDetail> getMyPointDetail(Long userId);
+    public PointDetail chargePoint(Integer userId, Long point);
+    public PointDetail usePoint(Integer userId, Long point);
+    public Long getMyPoint(Integer userId);
+    public List<PointDetail> getMyPointDetail(Integer userId);
 
-    public void payForReview(Long userId, Long reviewId,Long point);
-    public boolean isAlreadyPaid(Long userId, Long reviewId);
+    public void payForReview(Integer userId, Integer reviewId,Long point);
+    public boolean isAlreadyPaid(Integer userId, Integer reviewId);
 
 }
