+++ src/main/java/com/example/block/service/ReviewService.java
@@ -3,6 +3,6 @@ package com.example.block.service;
 import com.example.block.domain.mapping.Review;
 
 public interface ReviewService {
-    public Review getReviewDetail(Long userId,Long reviewId);
+    public Review getReviewDetail(Integer userId,Integer reviewId);
 
 }
